server/interview_agent.py
import os
import io
import json
from typing import List, Dict
import PyPDF2
import docx
from openai import OpenAI
from elevenlabs import VoiceSettings
from elevenlabs.client import ElevenLabs
from dotenv import load_dotenv

load_dotenv()

# Initialize clients lazily or with validation
def get_openai_client():
    api_key = os.getenv("OPENAI_API_KEY")
    if not api_key or "your_openai_api_key" in api_key:
        raise ValueError("Missing or invalid OPENAI_API_KEY in .env file")
    return OpenAI(api_key=api_key)

def get_eleven_client():
    api_key = os.getenv("ELEVENLABS_API_KEY")
    if not api_key or "your_elevenlabs_api_key" in api_key:
        raise ValueError("Missing or invalid ELEVENLABS_API_KEY in .env file")
    print(f"Initializing ElevenLabs client with key starting with: {api_key[:5]}...")
    return ElevenLabs(api_key=api_key)

def extract_text_from_pdf(file_bytes):
    pdf_reader = PyPDF2.PdfReader(io.BytesIO(file_bytes))
    text = ""
    for page in pdf_reader.pages:
        text += page.extract_text()
    return text

def extract_text_from_docx(file_bytes):
    doc = docx.Document(io.BytesIO(file_bytes))
    text = "\n".join([para.text for para in doc.paragraphs])
    return text

import random

def generate_interview_questions(resume_text: str) -> List[str]:
    print(f"Generating questions for resume text length: {len(resume_text)}")
    
    # 15 General Interview Questions
    question_pool = [
        "What are your major skills?",
        "Tell me about your best project.",
        "What are your strengths?",
        "What are your weaknesses?",
        "Why do you want to join our company?",
        "Where do you see yourself in five years?",
        "How do you handle pressure and stress?",
        "Tell me about a time you faced a conflict at work and how you resolved it.",
        "What is your greatest professional achievement?",
        "How do you stay updated with the latest trends in your field?",
        "What do you look for in a team environment?",
        "Describe a situation where you had to learn a new technology quickly.",
        "What are your salary expectations?",
        "Do you have any questions for us?",
        "What motivates you to do your best work?"
    ]

    # Randomly pick 4 more questions from the remaining 14
    random_questions = random.sample(question_pool, 4)
    
    # The first question is always "Tell me about yourself."
    final_questions = ["Hello! I've reviewed your resume. Let's start. Tell me about yourself."] + random_questions
    
    print(f"Questions selected: {final_questions}")
    return final_questions

def get_follow_up_question(resume_text: str, conversation_history: List[Dict[str, str]]) -> str:
    print(f"Generating follow-up question. History length: {len(conversation_history)}")
    try:
        messages = [
            {"role": "system", "content": f"You are an expert technical interviewer. Use the resume for context: {resume_text}"},
        ]
        messages.extend(conversation_history)
        messages.append({"role": "user", "content": "Generate a dynamic follow-up question based on the last answer."})
        
        client = get_openai_client()
        response = client.chat.completions.create(
            model="gpt-4o",
            messages=messages
        )
        
        return response.choices[0].message.content
    except Exception as e:
        print(f"Error generating follow-up question: {e}")
        # Fallback if AI fails to generate follow-up
        return "Can you tell me more about that?"

def text_to_speech(text: str) -> bytes:
    print(f"Generating TTS for: {text[:30]}...")
    voice_id = os.getenv("ELEVENLABS_VOICE_ID")
    if not voice_id:
        voice_id = "pNInz6obpg8nEmeWscDJ" # Default: Rachel
    
    try:
        # Try ElevenLabs first
        client = get_eleven_client()
        response = client.text_to_speech.convert(
            voice_id=voice_id,
            output_format="mp3_44100_128",
            text=text,
            model_id="eleven_multilingual_v2",
            voice_settings=VoiceSettings(
                stability=0.5,
                similarity_boost=0.75,
                style=0.0,
                use_speaker_boost=True,
            ),
        )
        # ElevenLabs returns a generator of bytes. We must consume it here.
        audio_data = b"".join(response)
        print(f"ElevenLabs TTS successful. Size: {len(audio_data)} bytes")
        return audio_data
    except Exception as e:
        print(f"ElevenLabs TTS Error: {e}. Falling back to OpenAI TTS.")
        try:
            # Fallback to OpenAI TTS
            client_oa = get_openai_client()
            response_oa = client_oa.audio.speech.create(
                model="tts-1",
                voice="alloy",
                input=text
            )
            print(f"OpenAI TTS fallback successful. Size: {len(response_oa.content)} bytes")
            return response_oa.content
        except Exception as oa_e:
            print(f"OpenAI TTS Fallback Error: {oa_e}")
            raise oa_e

def speech_to_text(audio_file_path: str) -> str:
    client = get_openai_client()
    with open(audio_file_path, "rb") as audio_file:
        transcript = client.audio.transcriptions.create(
            model="whisper-1", 
            file=audio_file
        )
    return transcript.text

def evaluate_interview(resume_text: str, q_and_a: List[Dict[str, str]]) -> Dict:
    print("Evaluating interview content...")
    qa_summary = "\n".join([f"Q: {item['question']}\nA: {item['answer']}" for item in q_and_a])
    
    prompt = f"""
    Evaluate this interview based on the candidate's resume and their answers.
    Provide scores (0-100) for Clarity, Confidence, and Fluency.
    Also provide a list of feedback points.
    
    Resume:
    {resume_text}
    
    Interview Q&A:
    {qa_summary}
    
    Return JSON format:
    {{
        "clarity": number,
        "confidence": number,
        "fluency": number,
        "feedback": ["point 1", "point 2", ...]
    }}
    """
    
    try:
        client = get_openai_client()
        response = client.chat.completions.create(
            model="gpt-4o",
            messages=[{"role": "user", "content": prompt}],
            response_format={"type": "json_object"}
        )
        return json.loads(response.choices[0].message.content)
    except Exception as e:
        print(f"Evaluation AI Error (Quota?): {e}")
        # Return fallback scores so report is not empty
        return {
            "clarity": 70,
            "confidence": 75,
            "fluency": 72,
            "feedback": [
                "Good attempt at answering questions.",
                "Note: Detailed content analysis was unavailable due to service quota.",
                "Focus on providing more structured answers in the next session."
            ]
        }
