server/requirements.txt
fastapi
uvicorn
python-multipart
librosa
numpy
scipy
soundfile
openai
elevenlabs
python-docx
PyPDF2
python-dotenv
